# specs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Customer Service Demo API
  version: 1.0.0
  description: >
    Frozen V1 API draft for the Web-first e-commerce AI customer service demo.
    All endpoints are scoped to an anonymous Demo Workspace.
servers:
  - url: /api
security:
  - DemoWorkspaceToken: []
...
paths:
...
  /workflows:
    parameters:
      - in: query
        name: shopId
        required: false
        description: Optional shop scope for shop-bound workflow definitions.
        schema: { type: string }
    get:
      tags: [Workflows]
      summary: List workflows
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      tags: [Workflows]
      summary: Create a workflow draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:
                  type: string
                type:
                  type: string
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    get:
      tags: [Workflows]
      summary: Get one workflow and its current draft/active version
      responses:
        "200":
          description: Workflow snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          $ref: "#/components/responses/NotFound"
  /workflows/{workflowId}/draft:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    put:
      tags: [Workflows]
      summary: Save workflow draft graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowGraph"
      responses:
        "200":
          description: Draft saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}/publish:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    post:
      tags: [Workflows]
      summary: Validate and publish a new immutable workflow version
      responses:
        "201":
          description: Published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
  /workflows/{workflowId}/enable:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    post:
      tags: [Workflows]
      summary: Enable a published workflow
      responses:
        "202":
          description: Enable operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /workflows/{workflowId}/disable:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    post:
      tags: [Workflows]
      summary: Disable a workflow
      responses:
        "202":
          description: Disable operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /workflows/{workflowId}/test-run:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"
    post:
      tags: [Workflows]
      summary: Run workflow against synthetic context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId]
              properties:
                conversationId:
                  type: string
      responses:
        "202":
          description: Test run accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /workflow-runs:
    parameters:
      - in: query
        name: workflowId
        required: false
        schema: { type: string }
      - in: query
        name: conversationId
        required: false
        schema: { type: string }
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum: [PENDING, RUNNING, WAITING_APPROVAL, RECOVERING, COMPLETED, FAILED, STALE, CANCELLED]
    get:
      tags: [Workflows]
      summary: List workflow runs
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowRun"
  /workflow-runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Workflows]
      summary: Get one immutable workflow run with node runs
      responses:
        "200":
          description: Workflow run snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /action-proposals/{proposalId}/approve:
    parameters:
      - $ref: "#/components/parameters/ProposalId"
    post:
      tags: [Workflows]
      summary: Approve proposal, revalidate, then execute
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveActionProposalInput"
      responses:
        "202":
          description: Approval accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
  /action-proposals/{proposalId}/reject:
    parameters:
      - $ref: "#/components/parameters/ProposalId"
    post:
      tags: [Workflows]
      summary: Reject proposal
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectActionProposalInput"
      responses:
        "202":
          description: Rejection operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
...
components:
  securitySchemes:
    DemoWorkspaceToken:
      type: apiKey
      in: header
      name: X-Demo-Workspace-Token
  responses:
    NotFound:
      description: Resource not found in the current Demo Workspace
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  parameters:
...
    WorkflowId:
      in: path
      name: workflowId
      required: true
      schema: { type: string }
    RunId:
      in: path
      name: runId
      required: true
      schema: { type: string }
    ProposalId:
      in: path
      name: proposalId
      required: true
      schema: { type: string }
...
  schemas:
...
    Workflow:
      type: object
      required: [id, name, type, priority, status]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        tenantId: { type: string }
        shopId: { type: [string, 'null'] }
        name: { type: string }
        type: { type: string }
        priority: { type: integer, minimum: 0 }
        status: { type: string, enum: [DRAFT, PUBLISHED, DISABLED] }
        activeVersionId: { type: [string, 'null'] }
        draftVersion:
          oneOf:
            - $ref: "#/components/schemas/WorkflowVersion"
            - type: 'null'
        activeVersion:
          oneOf:
            - $ref: "#/components/schemas/WorkflowVersion"
            - type: 'null'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    WorkflowVersion:
      type: object
      required: [id, workflowId, version, graph]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        tenantId: { type: string }
        workflowId: { type: string }
        version: { type: integer, minimum: 0 }
        graph:
          $ref: "#/components/schemas/WorkflowGraph"
        publishedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    WorkflowSettings:
      type: object
      required: [maxSteps, timeoutMs]
      properties:
        maxSteps: { type: integer, minimum: 1, maximum: 20 }
        timeoutMs: { type: integer, minimum: 1, maximum: 30000 }
    WorkflowGraph:
      type: object
      required: [nodes, edges, settings]
      properties:
        nodes:
          type: array
          items:
            type: object
            required: [id, type, position, config]
            properties:
              id: { type: string, minLength: 1 }
              type:
                type: string
                enum:
                  [TRIGGER, CONDITION, QUERY_PRODUCT, QUERY_ORDER, QUERY_LOGISTICS, AI_GENERATE, HUMAN_APPROVAL, END]
              position:
                type: object
                required: [x, y]
                properties:
                  x: { type: number }
                  y: { type: number }
              config:
                type: object
                additionalProperties: true
        edges:
          type: array
          items:
            type: object
            required: [id, source, target]
            properties:
              id: { type: string, minLength: 1 }
              source: { type: string, minLength: 1 }
              target: { type: string, minLength: 1 }
              condition: { type: string, enum: ["true", "false"] }
        settings:
          $ref: "#/components/schemas/WorkflowSettings"
    NodeRun:
      type: object
      required: [id, workflowRunId, nodeId, status, retryCount]
      properties:
        id: { type: string }
        workflowRunId: { type: string }
        nodeId: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED, WAITING_APPROVAL, STALE, SKIPPED] }
        input: { type: [object, 'null'] }
        output: { type: [object, 'null'] }
        errorCode: { type: [string, 'null'] }
        retryCount: { type: integer, minimum: 0 }
        startedAt: { type: [string, 'null'], format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        durationMs: { type: [integer, 'null'], minimum: 0 }
    WorkflowRun:
      type: object
      required: [id, shopId, conversationId, workflowVersionId, taskIds, contextVersion, completedNodeIds, status, startedAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        tenantId: { type: string }
        shopId: { type: string }
        conversationId: { type: string }
        workflowVersionId: { type: string }
        taskIds:
          type: array
          items: { type: string }
        contextVersion: { type: integer, minimum: 0 }
        currentNodeId: { type: [string, 'null'] }
        completedNodeIds:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [PENDING, RUNNING, WAITING_APPROVAL, RECOVERING, COMPLETED, FAILED, STALE, CANCELLED]
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
        nodeRuns:
          type: array
          items:
            $ref: "#/components/schemas/NodeRun"
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/ActionProposal"
...
    ActionProposal:
      type: object
      required: [id, shopId, conversationId, type, riskLevel, targetEntityType, targetEntityId, contextVersion, status]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        tenantId: { type: string }
        shopId: { type: string }
        conversationId: { type: string }
        workflowRunId: { type: [string, 'null'] }
        type: { type: string, enum: [MARK_READ, CREATE_INTERNAL_TASK, TRANSFER_HUMAN, ADD_ORDER_REMARK, PROPOSE_COMPENSATION, REFUND, EXCHANGE] }
        riskLevel: { type: string, enum: [READ, LOW_WRITE, MEDIUM_WRITE, HIGH_RISK] }
        targetEntityType: { type: string }
        targetEntityId: { type: string }
        payload: { type: object, additionalProperties: true }
        evidenceIds:
          type: array
          items: { type: string }
        contextVersion: { type: integer, minimum: 0 }
        status: { type: string, enum: [PROPOSED, POLICY_CHECKED, WAITING_APPROVAL, APPROVED, REVALIDATING, EXECUTING, SUCCEEDED, REJECTED, STALE, FAILED, UNCERTAIN, CANCELLED] }
        approvedBy: { type: [string, 'null'] }
        approvedAt: { type: [string, 'null'], format: date-time }
        receipt: { type: [object, 'null'], additionalProperties: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApproveActionProposalInput:
      type: object
      properties:
        expectedContextVersion: { type: integer, minimum: 0 }
        idempotencyKey: { type: string, minLength: 1 }
    RejectActionProposalInput:
      type: object
      properties:
        reason: { type: string }
...
    OperationAccepted:
      type: object
      required: [operationId, status]
      properties:
        operationId: { type: string }
        status: { type: string, enum: [ACCEPTED, QUEUED] }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }